Move decryption to MavenSecuritySettings
